From Chet Ramey:
Official patches 3, 4, and 5 to bash-1.13, plus a new FreeBSD
machine description and a couple of fixes from bash-1.14.

+++ shell/bash/getcwd.c
@@ -108,7 +108,7 @@ extern char *alloca ();
    big as necessary.  */
 #if defined (__STDC__)
 char *
-getcwd (char *buf, int size)
+getcwd (char *buf, size_t size)
 #else /* !__STDC__ */
 char *
 getcwd (buf, size)
